Fixed userscript not loading in a few cases
Youtube being a single page app disguised as a multi page app made it so just matching the url wasn't enough for loading the userscript as it didn't trigger a check from scriptrunners like Tampermonkey. In other words the userscript has to always be running anywhere on Youtube, otherwise it won't work.